requirements.txt
streamlit
PyMuPDF
faiss-cpu
numpy
langchain
langchain-google-genai
google-generativeai
